created separate classes file
